# docs/swagger/swagger.yaml
definitions:
  models.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: This is an API to get item details for MercadoLibre frontend page.
  termsOfService: http://swagger.io/terms/
  title: MELI Item Detail
  version: "1.0"
paths:
  /api/v1/item/{itemID}:
    get:
      description: Getting information for provided item
      parameters:
      - description: The item id
        in: path
        name: itemID
        required: true
        type: string
      - description: User id
        in: query
        name: userID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: The specified URL is invalid
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.Response'
        "408":
          description: Request timed out
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get item info
      tags:
      - item
  /api/v1/recommendations/{itemID}:
    get:
      description: Get recommendations for a given user and item
      parameters:
      - description: The item id
        in: path
        name: itemID
        required: true
        type: string
      - description: Seller id
        in: query
        name: sellerID
        required: true
        type: string
      - description: User id
        in: query
        name: userID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: The specified URL is invalid
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.Response'
        "408":
          description: Request timed out
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get recommendations
      tags:
      - item
swagger: "2.0"